+++ tests/testthat/test-ranks.R
@@ -6,16 +6,17 @@ lineage3 <- paste0("k__Bacteria|p__Proteobacteria|c__Betaproteobacteria",
                    "|o__Burkholderiales|f__Comamonadaceae",
                    "|g__Delftia|s__Delftia_unclassified")
 lineage4 <- "k__Bac|p__Fir|c__Clos|o__Clost|f__Rumi|g__Subdo|s__Sub_su|t__X_56Z"
-lineages <- c(lineage1, lineage2, lineage3, lineage4)
+lineage5 <- "k__Viruses"
+lineages <- c(lineage1, lineage2, lineage3, lineage4, lineage5)
 
 
 ## is_rank()
 
 test_that("is_rank() is correct", {
-  expect_equal(is_rank(lineages, "class"), c(TRUE, TRUE, FALSE, FALSE))
-  expect_equal(is_rank(lineages, "order"), c(FALSE, FALSE, FALSE, FALSE))
-  expect_equal(is_rank(lineages, "sp"), c(FALSE, FALSE, TRUE, FALSE))
-  expect_equal(is_rank(lineages, "strain"), c(FALSE, FALSE, FALSE, TRUE))
+  expect_equal(is_rank(lineages, "class"), c(TRUE, TRUE, FALSE, FALSE, FALSE))
+  expect_equal(is_rank(lineages, "order"), c(FALSE, FALSE, FALSE, FALSE, FALSE))
+  expect_equal(is_rank(lineages, "sp"), c(FALSE, FALSE, TRUE, FALSE, FALSE))
+  expect_equal(is_rank(lineages, "strain"), c(FALSE, FALSE, FALSE, TRUE, FALSE))
 })
 
 test_that("is_rank() throws error when needed", {
@@ -27,13 +28,13 @@ test_that("is_rank() throws error when needed", {
 
 test_that("is_clade() is correct", {
   expect_equal(is_clade(lineages, "Clostridia", "class"),
-               c(FALSE, TRUE, FALSE, FALSE))
+               c(FALSE, TRUE, FALSE, FALSE, FALSE))
   expect_equal(is_clade(lineages, "Verrucomicrobia"),
-               c(TRUE, FALSE, FALSE, FALSE))
+               c(TRUE, FALSE, FALSE, FALSE, FALSE))
   expect_equal(is_clade(lineages, "Bacteria"),
-               c(TRUE, TRUE, TRUE, FALSE))
+               c(TRUE, TRUE, TRUE, FALSE, FALSE))
   expect_equal(is_clade(lineages, "Bacter"),
-               c(FALSE, FALSE, FALSE, FALSE))
+               c(FALSE, FALSE, FALSE, FALSE, FALSE))
 })
 
 test_that("is_clade() throws error when needed", {
@@ -43,14 +44,28 @@ test_that("is_clade() throws error when needed", {
 })
 
 
+## last_clade()
+
+test_that("last_clade() is correct", {
+  expect_equal(last_clade(lineages[1:2]),
+               c("Verrucomicrobiae", "Clostridia"))
+  expect_equal(last_clade(lineages, same = FALSE),
+               c("Verrucomicrobiae", "Clostridia",
+                 "Delftia_unclassified", "X_56Z", "Viruses"))
+})
+
+test_that("last_clade() throws error when needed", {
+  expect_error(last_clade(lineages, same = TRUE))
+})
+
+
 ## last_rank()
 
 test_that("last_rank() is correct", {
   expect_equal(last_rank(lineages[1:2]),
-               c("Verrucomicrobiae", "Clostridia"))
+               c("class", "class"))
   expect_equal(last_rank(lineages, same = FALSE),
-               c("Verrucomicrobiae", "Clostridia",
-                 "Delftia_unclassified", "X_56Z"))
+               c("class", "class", "species", "strain", "kingdom"))
 })
 
 test_that("last_rank() throws error when needed", {