+++ NEWS.md
@@ -1,6 +1,11 @@
+# yatah 0.0.0.9004
+
+* Changed `last_rank()` into `last_clade` (breaking change).
+* New function `last_rank()`.
+
 # yatah 0.0.0.9003
 
-* Used `rank` instead of `level` (breaking change)
+* Used `rank` instead of `level` (breaking change).
 
 # yatah 0.0.0.9002